(xstack-intel) OCR: Merge MPI branch into runtime/master
Issues fixed:
  - Shutdown protocol was screwed up in the merge
  - Re-added Single Assignment DBs as some of the tests still
    use it (is it still relevant, maybe remove in a follow-up
    change)
  - Initialization of the allocators in TGEmul
  - Other minor things
  - db containing mainEdt arguments must be released by XE
    do not use ocrDbRelease to avoid WARN message interpreted
    by jenkins as a failure (release outside EDT)
  - Add code in CE PD to manage a busy DB acquire request/response
  - x86-pthread-x86 change default CFG file to a lockableDB one

[Closes issue #137]
[Closes issue #163]
[References issue #195]

Conflicts resolved:
	ocr/build/common.mk
	ocr/build/tg-null-tg-xe/Makefile
	ocr/ocr-apps/fib/fib.c
	ocr/runtime/ocr-x86/src/api/ocr.c
	ocr/runtime/ocr-x86/src/comm-platform/comm-platform-all.c
	ocr/runtime/ocr-x86/src/inc/ocr-runtime-types.h
	ocr/runtime/ocr-x86/src/policy-domain/ce/ce-policy.c
	ocr/runtime/ocr-x86/src/policy-domain/hc/hc-policy.c
	ocr/runtime/ocr-x86/src/task/hc/hc-task.c
	ocr/runtime/ocr-x86/src/worker/xe/xe-worker.c

+++ runtime/ocr-x86/src/policy-domain/ce/ce-policy.h
@@ -26,8 +26,8 @@ typedef struct {
 typedef struct {
     ocrPolicyDomain_t base;
     u8 xeCount;                     //number of XE's serviced by this CE
-    u8 shutdownCount;               //dynamic counter to keep track of agents alreadt shutdown
-    u8 shutdownMax;                 //maximum number of agents participating in the shutdown process
+    u32 shutdownCount;              //dynamic counter to keep track of agents alreadt shutdown
+    u32 shutdownMax;                //maximum number of agents participating in the shutdown process
     bool shutdownMode;              //flag indicating shutdown mode
     bool *ceCommTakeActive;         //flag to coordinate the search for work
 } ocrPolicyDomainCe_t;
